requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
appnope==0.1.4
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
backcall==0.2.0
base58==2.1.1
bitarray==2.9.2
boto3==1.29.1
botocore==1.32.1
Bottleneck==1.3.5
Brotli==1.0.9
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
comm==0.2.1
contourpy==1.2.0
cryptography==42.0.5
cycler==0.12.1
cytoolz==0.12.1
dateparser==1.0.0
debugpy==1.6.7
decorator==5.1.1
dydx-v3-python==2.1.0
ecdsa==0.18.0
entrypoints==0.4
eth-abi==2.2.0
eth-account==0.5.9
eth-hash==0.6.0
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.9.5
executing==2.0.1
fonttools==4.49.0
frozenlist==1.4.1
hexbytes==0.3.1
idna==3.4
importlib_resources==6.1.2
ipfshttpclient==0.8.0a2
ipykernel==6.29.2
ipython==8.12.0
jedi==0.19.1
jmespath==1.0.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter-client==7.3.4
jupyter_core==5.5.0
kiwisolver==1.4.5
lru-dict==1.3.0
matplotlib==3.8.3
matplotlib-inline==0.1.6
mpmath==1.0.0
multiaddr==0.0.9
multidict==6.0.5
nest_asyncio==1.6.0
netaddr==0.10.1
numexpr==2.8.7
numpy==1.26.3
packaging==23.1
pandas==2.1.4
parsimonious==0.8.1
parso==0.8.3
patsy==0.5.3
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.2.0
pip==23.3.1
platformdirs==4.2.0
prompt-toolkit==3.0.42
protobuf==3.19.5
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pycryptodome==3.20.0
Pygments==2.17.2
pyOpenSSL==24.0.0
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3.post1
pyzmq==25.1.2
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
rlp==2.0.1
rpds-py==0.17.1
s3transfer==0.7.0
scipy==1.11.4
setuptools==68.2.2
six==1.16.0
stack-data==0.6.2
statsmodels==0.14.0
sympy==1.6
toolz==0.12.1
tornado==6.1
traitlets==5.14.1
typing_extensions==4.9.0
tzdata==2023.3
tzlocal==5.2
urllib3==1.26.18
varint==1.0.2
wcwidth==0.2.13
web3==5.31.4
websockets==9.1
wheel==0.41.2
yarl==1.9.4
zipp==3.17.0
